// src/service/file/schemas/smartFolderItemResultSmartFolderApiType.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * File API
 * isd / kinder board api doc
 * OpenAPI spec version: v1.1.0
 */

/**
 * api 요청하는 구분(Photo:사진스마트폴더, EducationalData:자료스마트폴더, UserFolder:내폴더)
 */
export type SmartFolderItemResultSmartFolderApiType =
  (typeof SmartFolderItemResultSmartFolderApiType)[keyof typeof SmartFolderItemResultSmartFolderApiType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SmartFolderItemResultSmartFolderApiType = {
  Photo: 'Photo',
  EducationalData: 'EducationalData',
  UserFolder: 'UserFolder',
  MyBoard: 'MyBoard',
  PublicItem: 'PublicItem',
  Reply: 'Reply',
  Trashcan: 'Trashcan',
  Drive: 'Drive',
} as const;